refactor: 화면 좌표를 정규화 대신 픽셀로 싣는다
정규화 값은 화면 크기를 버린다. 픽셀과 해상도를 함께 실으면 소비자가 나누기
한 번으로 정규화 값을 얻지만, 반대는 복원할 수 없다. 같은 비용에 정보를 줄일
이유가 없어 rect 를 픽셀로 바꾸고 scene 에 screen 을 동봉한다.

양자화도 이쪽이 자연스럽다. 정규화 4자리는 1080p 에서 약 0.2px 이라 미세
애니메이션이 해시를 흔들 여지가 남았는데, 정수 픽셀은 반올림 자체가 1px 양자다.

대가는 스트림이 다운스케일된 경우 소비자가 frame.w / screen.w 로 환산해야
한다는 것이고, 명세에 적었다.

Jira: ARTEL-153

+++ Packages/kr.artel.sdk/Runtime/Protocol/Dto/ScreenRectDto.cs
@@ -3,25 +3,25 @@ using Newtonsoft.Json;
 namespace Artel.Protocol.Dto
 {
     /// <summary>
-    /// The area a block covers on screen, normalized to 0..1 with the origin at the top left.
+    /// The area a block covers on screen, in pixels, with the origin at the top left.
     /// </summary>
     /// <remarks>
-    /// <see cref="X"/> and <see cref="Y"/> are the top-left corner rather than the centre.
-    /// <see cref="W"/> is divided by the screen's width and <see cref="H"/> by its height, so equal
-    /// values do not describe a square.
+    /// <see cref="X"/> and <see cref="Y"/> are the top-left corner rather than the centre. The
+    /// pixels are the game's own screen, which the scene reports alongside as <c>screen</c>; a
+    /// reader working against a downscaled video frame scales by the ratio between the two.
     /// </remarks>
     public sealed class ScreenRectDto
     {
         [JsonProperty("x")]
-        public float X { get; set; }
+        public int X { get; set; }
 
         [JsonProperty("y")]
-        public float Y { get; set; }
+        public int Y { get; set; }
 
         [JsonProperty("w")]
-        public float W { get; set; }
+        public int W { get; set; }
 
         [JsonProperty("h")]
-        public float H { get; set; }
+        public int H { get; set; }
     }
 }
